requirements.txt
python-dotenv==1.1.0
langchain-community==0.3.22
fastapi==0.115.12
wikipedia==1.4.0
pypdf==5.4.0
unstructured==0.17.2
unstructured[pdf]
langchain-pinecone==0.2.5
langchain-openai==0.3.14
langchain-unstructured==0.1.6
serpapi==0.1.5
google_search_results==2.4.2
langgraph==0.4.1
ipython==9.2.0
jupyterlab==4.4.1
playwright==1.52.0
beautifulsoup4==4.13.4
html2text==2025.4.15
langgraph==0.4.1
pytesseract==0.3.13
pdf2image==1.17.0
transformers==4.52.4
sentence-transformers==4.1.0
faiss-cpu==1.11.0
PyMuPDF==1.26.1
easyocr==1.7.2
sentence-transformers==4.1.0
accelerate==1.7.0
bitsandbytes==0.46.0
uvicorn==0.34.3
streamlit==1.45.1
